+++ examples/render_single_template.rs
@@ -1,15 +1,14 @@
 use tera::Context;
-use terarium::{Template, TerariumBuilder};
+use terarium::{Content, Template, TerariumBuilder};
 
 /// The Terarium can render single template.
 fn main() {
     let terarium = TerariumBuilder::default()
         .add_template(
             "my_template".to_owned(),
-            Template::default().content_builder()
-                .add_content("This is my template #{{tpl_number}}".to_owned(), vec!["en".to_owned()])
-                .add_content("Toto je šablona #{{tpl_number}}".to_owned(), vec!["cs".to_owned()])
-                .build()
+            Template::default()
+                .add_content(Content::new("This is my template #{{tpl_number}}".to_owned(), vec!["en".to_owned()])).unwrap()
+                .add_content(Content::new("Toto je šablona #{{tpl_number}}".to_owned(), vec!["cs".to_owned()])).unwrap()
         )
         .build().unwrap();